"Working" Parser
The parser is running and forming a Json object. However there are a few
bugs I still need to work out. Some of these bugs include:

Not recording the winner
incorrectly naming things with a \n
adding in new entries to the entry hazard dictionary
Not recording Pokemon moves (maybe) this is a maybe because the pokemon
may have been KOed before a move could be used.