Trying to add in random moves
I've been at this all day. I don't have a working database so the
program will not run.

Here's what I did:

query(movest).filter(Pokemon.name == pkmn['name']).all()
I'm hoping that this will actually return a list of all the moves this
pokemon can learn. I then generate a random number and check if the move
that I generate is already in the move list. If so, try again, else
append it to the move list.

This should work as the moves we append don't really matter, but it
gives difference to the pokemon.